(Refactor, Docs) 코드 개선, 주석 작성
일부 for문 자료형 수정
인덱스 반환 메소드 null 반환으로 수정
일부 코드 가독성 개선
To do 주석, 설명 주석 작성

+++ Assets/Scripts/Inventory/InventoryUI.cs
@@ -39,7 +39,7 @@ public class InventoryUI : MonoBehaviour
     public void UpdateInventory()
     {
         Debug.Log("Udate Inventory", gameObject);
-        for (int i = 0; i < slots.Length; i++)
+        for (byte i = 0; i < slots.Length; i++)
         {
             slots[i].ApplyInventoryData(inventory.inventory[i]);
         }
